[XLA:GPU][NFC] Consolidate CuTeDSL module implementation
Move ModuleImage into module.{h,cc} and remove the redundant source files and Bazel target.
