+++ xla/backends/gpu/libraries/cutedsl/collective_ffi.cc
@@ -40,7 +40,6 @@ limitations under the License.
 #include "xla/backends/gpu/libraries/cutedsl/config.h"
 #include "xla/backends/gpu/libraries/cutedsl/ffi_abi.h"
 #include "xla/backends/gpu/libraries/cutedsl/module.h"
-#include "xla/backends/gpu/libraries/cutedsl/module_image.h"
 #include "xla/backends/gpu/runtime/collective_clique_requests.h"
 #include "xla/backends/gpu/runtime/collective_cliques.h"
 #include "xla/backends/gpu/runtime/collective_execution.h"
